Fix new clippy lints (#832)
* fix clippy

* fmt

* revert taplofmt

* Revert "fmt"

This reverts commit 8fe90f08f1b70f3e31a451a588bc38dbade73280.

* cargo fmt

+++ zrml/prediction-markets/src/benchmarks.rs
@@ -490,7 +490,7 @@ benchmarks! {
     }
 
     dispute_authorized {
-        let d in 0..(T::MaxDisputes::get() - 1) as u32;
+        let d in 0..(T::MaxDisputes::get() - 1);
         let b in 0..63;
 
         let report_outcome = OutcomeReport::Scalar(u128::MAX);
